ocr_service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
easyocr==1.7.1
pillow==9.5.0 